refactor: extract table row to own component
