+++ packages/dm-core-plugins/src/table/TablePlugin.tsx
@@ -1,4 +1,4 @@
-import React, { ChangeEvent, useEffect, useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import {
   ErrorResponse,
   IUIPlugin,
@@ -6,38 +6,27 @@ import {
   Pagination,
   Stack,
   TGenericObject,
-  ViewCreator,
   useDMSS,
   useDocument,
 } from '@development-framework/dm-core'
-import {
-  Button,
-  EdsProvider,
-  Icon,
-  Input,
-  Table,
-  Tooltip,
-} from '@equinor/eds-core-react'
-import {
-  add,
-  chevron_down,
-  chevron_up,
-  delete_to_trash,
-  minimize,
-} from '@equinor/eds-icons'
+import { Button, Icon, Table } from '@equinor/eds-core-react'
+import { add } from '@equinor/eds-icons'
 import { AxiosError, AxiosResponse } from 'axios'
 import { SaveButton } from '../generic-list/Components'
-import { moveItem } from '../generic-list/utils'
-import { TTableItemRow, TTablePluginConfig, defaultConfig } from './types'
+import { TTableRowItem, TTablePluginConfig, defaultConfig } from './types'
+import { TableRow } from './TableRow/TableRow'
 
 export const TablePlugin = (props: IUIPlugin) => {
   const { idReference, type, onOpen = () => null } = props
-  const config: TTablePluginConfig = { ...defaultConfig, ...props.config }
-  const functionality = {
-    ...defaultConfig.functionality,
-    ...props.config.functionality,
+  const config: TTablePluginConfig = {
+    ...defaultConfig,
+    ...props.config,
+    functionality: {
+      ...defaultConfig.functionality,
+      ...props.config?.functionality,
+    },
   }
-  const [items, setItems] = useState<TTableItemRow[]>([])
+  const [items, setItems] = useState<TTableRowItem[]>([])
   const [isSaveLoading, setIsSaveLoading] = useState<boolean>(false)
   const [dirtyState, setDirtyState] = useState<boolean>(false)
   const [paginationPage, setPaginationPage] = useState<number>(0)
@@ -74,14 +63,6 @@ export const TablePlugin = (props: IUIPlugin) => {
     setItems(itemsWithIds)
   }, [document, loading])
 
-  function deleteItem(reference: string, key: string) {
-    const itemIndex = items.findIndex((item) => item.key === key)
-    const itemsCopy = [...items]
-    itemsCopy.splice(itemIndex, 1)
-    setItems(itemsCopy)
-    setDirtyState(true)
-  }
-
   function addItem() {
     dmssAPI
       .instantiateEntity({
@@ -105,18 +86,6 @@ export const TablePlugin = (props: IUIPlugin) => {
       )
   }
 
-  function updateItem(
-    index: number,
-    attribute: string,
-    newValue: string | number | boolean
-  ) {
-    const itemsCopy = [...items]
-    itemsCopy[index].data[attribute] = newValue
-    itemsCopy[index].isSaved = false
-    setItems(itemsCopy)
-    setDirtyState(true)
-  }
-
   function saveTable() {
     setIsSaveLoading(true)
     const payload = items.map((item) => item.data)
@@ -126,7 +95,7 @@ export const TablePlugin = (props: IUIPlugin) => {
         data: JSON.stringify(payload),
       })
       .then(() => {
-        const updatedItems: TTableItemRow[] = items.map((item) => {
+        const updatedItems: TTableRowItem[] = items.map((item) => {
           return { ...item, isSaved: true }
         })
         setItems(updatedItems)
@@ -138,25 +107,6 @@ export const TablePlugin = (props: IUIPlugin) => {
       .finally(() => setIsSaveLoading(false))
   }
 
-  function openItemAsTab(item: TGenericObject, index: number) {
-    const view = { label: item?.data?.name, type: 'ViewConfig' }
-    onOpen(crypto.randomUUID(), view, `${idReference}[${index}]`)
-  }
-
-  function expandItem(index: number) {
-    const itemsCopy = [...items]
-    itemsCopy[index].expanded = !itemsCopy[index].expanded
-    setItems(itemsCopy)
-  }
-
-  function getColumnsLength() {
-    let amount = config.columns?.length || 0
-    if (functionality.delete) amount += 1
-    if (functionality.openAsTab || functionality.openAsExpandable) amount += 1
-    if (functionality.sort) amount += 1
-    return amount
-  }
-
   if (error) throw new Error(JSON.stringify(error, null, 2))
   if (loading) return <Loading />
 
@@ -165,141 +115,35 @@ export const TablePlugin = (props: IUIPlugin) => {
       <Table style={{ width: '100%' }}>
         <Table.Head>
           <Table.Row>
-            {(functionality?.openAsTab || functionality?.openAsExpandable) && (
+            {(config.functionality?.openAsTab ||
+              config.functionality?.openAsExpandable) && (
               <Table.Cell width="80"></Table.Cell>
             )}
             {config.columns.map((attribute: string) => (
               <Table.Cell key={attribute}>{attribute}</Table.Cell>
             ))}
-            {functionality?.delete && (
+            {config.functionality?.delete && (
               <Table.Cell width="48" aria-label="Delete"></Table.Cell>
             )}
-            {functionality?.sort && (
+            {config.functionality?.sort && (
               <Table.Cell width="48" aria-label="Sort"></Table.Cell>
             )}
           </Table.Row>
         </Table.Head>
         <Table.Body>
           {paginatedRows?.map((item, index) => (
-            <>
-              <Table.Row key={item.key}>
-                {(functionality?.openAsTab ||
-                  functionality.openAsExpandable) && (
-                  <Table.Cell>
-                    <Tooltip
-                      title={
-                        functionality.openAsExpandable
-                          ? item.expanded
-                            ? 'Minimize'
-                            : 'Expand'
-                          : 'Open in new tab'
-                      }
-                    >
-                      <Button
-                        variant="ghost_icon"
-                        color="secondary"
-                        disabled={!item.isSaved}
-                        onClick={
-                          functionality.openAsExpandable
-                            ? () => expandItem(index)
-                            : () => openItemAsTab(item, index)
-                        }
-                      >
-                        <Icon data={item.expanded ? minimize : add} />
-                      </Button>
-                    </Tooltip>
-                  </Table.Cell>
-                )}
-                {config.columns.map((attribute: string) => {
-                  if (typeof item.data[attribute] === 'object')
-                    throw new Error(
-                      `Objects can not be displayed in table. Attribute '${attribute}' is not a primitive type.`
-                    )
-                  // TODO: Consider having a more robust way of getting type and validating form
-                  const attributeType = typeof item.data[attribute]
-                  return (
-                    <Table.Cell key={attribute}>
-                      {functionality?.edit &&
-                      config.editableColumns?.includes(attribute) ? (
-                        <Input
-                          value={item.data[attribute] ?? ''}
-                          readOnly={attribute === 'type'}
-                          type={attributeType}
-                          onChange={(event: ChangeEvent<HTMLInputElement>) => {
-                            let newValue: string | number | boolean =
-                              event.target.value
-                            if (attributeType === 'number')
-                              newValue = Number(newValue)
-                            updateItem(index, attribute, newValue)
-                          }}
-                        />
-                      ) : (
-                        item.data?.[attribute]
-                      )}
-                    </Table.Cell>
-                  )
-                })}
-                <EdsProvider density="compact">
-                  {functionality?.delete && (
-                    <Table.Cell style={{ textAlign: 'center' }}>
-                      <Button
-                        title="Delete row"
-                        color="danger"
-                        variant="ghost_icon"
-                        onClick={() =>
-                          deleteItem(`${idReference}[${index}]`, item.key)
-                        }
-                      >
-                        <Icon data={delete_to_trash} aria-hidden />
-                      </Button>
-                    </Table.Cell>
-                  )}
-                  {functionality?.sort && (
-                    <Table.Cell style={{ width: '48px' }}>
-                      <>
-                        <Button
-                          title="Move row up"
-                          disabled={index === 0}
-                          variant="ghost_icon"
-                          onClick={() => {
-                            setItems(moveItem(items, item.key, 'up'))
-                            setDirtyState(true)
-                          }}
-                        >
-                          <Icon data={chevron_up} aria-hidden />
-                        </Button>
-                        <Button
-                          title="Move row down"
-                          disabled={
-                            index === rowsPerPage - 1 ||
-                            index === items?.length - 1
-                          }
-                          variant="ghost_icon"
-                          onClick={() => {
-                            setItems(moveItem(items, item.key, 'down'))
-                            setDirtyState(true)
-                          }}
-                        >
-                          <Icon data={chevron_down} aria-hidden />
-                        </Button>
-                      </>
-                    </Table.Cell>
-                  )}
-                </EdsProvider>
-              </Table.Row>
-              {item?.expanded && (
-                <Table.Row>
-                  <Table.Cell colSpan={getColumnsLength()}>
-                    <ViewCreator
-                      idReference={`${idReference}[${item.index}]`}
-                      viewConfig={
-                        config.views?.[item.index] ?? config.defaultView
-                      }
-                    />
-                  </Table.Cell>
-                </Table.Row>
-              )}
-            </>
+            <TableRow
+              key={item?.key}
+              config={config}
+              item={item}
+              index={index}
+              idReference={idReference}
+              items={items}
+              setItems={setItems}
+              setDirtyState={setDirtyState}
+              onOpen={onOpen}
+              rowsPerPage={rowsPerPage}
+            />
           ))}
         </Table.Body>
       </Table>
@@ -312,13 +156,13 @@ export const TablePlugin = (props: IUIPlugin) => {
           count={items?.length || 0}
         />
         <Stack direction="row" spacing={1} justifyContent="flex-end">
-          {functionality?.add && (
+          {config.functionality?.add && (
             <Button onClick={() => addItem()} variant="outlined">
               <Icon data={add} />
               Add row
             </Button>
           )}
-          {(functionality?.edit || functionality?.add) && (
+          {(config.functionality?.edit || config.functionality?.add) && (
             <SaveButton
               onClick={() => saveTable()}
               disabled={isSaveLoading || !dirtyState}
