+++ packages/dm-core-plugins/src/table/types.ts
@@ -1,18 +1,20 @@
 import { TGenericObject, TViewConfig } from '@development-framework/dm-core'
 
+export type TTableFunctionalityConfig = {
+  openAsTab: boolean
+  openAsExpandable: boolean
+  add: boolean
+  sort: boolean
+  edit: boolean
+  delete: boolean
+}
+
 export type TTablePluginConfig = {
   editMode: boolean
   columns: string[]
   showDelete: boolean
   editableColumns?: string[]
-  functionality: {
-    openAsTab: boolean
-    openAsExpandable: boolean
-    add: boolean
-    sort: boolean
-    edit: boolean
-    delete: boolean
-  }
+  functionality: TTableFunctionalityConfig
   defaultView: TViewConfig
   views: TViewConfig[]
 }
@@ -34,10 +36,22 @@ export const defaultConfig: TTablePluginConfig = {
   views: [],
 }
 
-export type TTableItemRow = {
+export type TTableRowItem = {
   key: string
   data: TGenericObject
   index: number
   expanded: boolean
   isSaved: boolean
+}
+
+export type TTableRow = {
+  config: TTablePluginConfig
+  item: TTableRowItem
+  index: number
+  idReference: string
+  items: TTableRowItem[]
+  setItems: React.Dispatch<React.SetStateAction<TTableRowItem[]>>
+  setDirtyState: React.Dispatch<React.SetStateAction<boolean>>
+  onOpen: any
+  rowsPerPage: number
 }
